Move tool.rs to tool_v1.rs
